requirements.txt
scipy
numpy
dlib
opencv2

